pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "torchsense"
version = "0.0.14"
authors = [
    {name = "Xingwei Wang", email = "wxwjkl123@gmail.com"},
]
description = "Torchsense is a library for sensor data processing with PyTorch"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "numpy",
    "einops",
    "h5py", "pesq",
    "torchaudio", "torchinfo",
    "torchmetrics[audio]",
    "lightning",

]

[project.urls]
"Homepage" = "https://github.com/xibrer/torchsense"


[tool.setuptools.packages.find]
include = ["torchsense*"]